move ExpectedSchemaOwner to owner package
